Autotriage.ai/requirements.txt
# Core dependencies
numpy>=1.26.0,<1.27.0
pandas>=2.1.4,<2.2.0
streamlit==1.29.0
google-generativeai==0.4.0
python-dotenv==1.0.0

# Visualization
plotly==5.18.0
plotly-express==0.4.1

# Machine Learning
scikit-learn==1.3.2

# Utils
colorama==0.4.6



